requirements.txt
django>=4.0,<5.0
djangorestframework>=3.12,<4.0
celery>=5.0,<6.0
redis>=4.0,<5.0
psycopg2>=2.9,<3.0
python-dotenv>=0.19.0,<1.0
gunicorn>=21.0
numpy==1.24.3
pandas==2.0.3
